pyproject.toml
[project]
name = "fari-digital-twin-backend-v2"
version = "0.1.0"
description = "FARI Digital Twin Backend V2"
authors = [
    { name = "Your Name", email = "you@example.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv",
    "requests",
    "pandas",
    "geopandas",
    "pyproj",
    "shapely",
    "beautifulsoup4",
    "numpy",
    "digitaltwin-dataspace",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["sources", "assets_manager"]

[tool.uv.sources]
digitaltwin-dataspace = { path = "../digitaltwin_dataspace" }
